python-apis/requirements.txt
flask
scikit-learn
joblib
pandas
numpy
flask-cors
requests
python-dotenv
google-generativeai>=0.3.2
numpy
openai==0.28